Trying to reduce CPU load during "idling"
It seems a lot of time is spent in this loop - 30% CPU load on my PI4. I don't see why you should spin loop here once you are connected - we only do something useful in those states, where we need to connect - once we are connected, we could actually drop out or at least wait for another state. But spinning for 1ms puts too much load on the CPU for an IOT device that's waiting for stuff to happen every few hours